Optionally fix balance errors
