+++ lib/double_entry/validation.rb
@@ -1,1 +1,2 @@
+require 'double_entry/validation/account_fixer'
 require 'double_entry/validation/line_check'
